Regenerate MFT SIA location when re-issuing TA certificate
When re-issuing a TA certificate, the SIA provided in the TA request are copied
onto the new TA certificate. However, this did not update the RPKI Manifest SIA
descriptor, as it is determined by TA-0 and therefore is not in the TA request.

This patch changes that by first updating the TA config URLs in the signing
context, then generate the default TA-0 SIA descriptors, and overwrite those
with the SIA descriptors in the TA request (repository location and RRDP
notification URL).